Event Visualization
A script for visualizing particle events was added. The script itself was created in a separate unity project, hence the lack of commits.
The particles and their textures are yet to be assigned in the scripts. Some textures seem to be missing from the original Belle II Leap SR project.